requirements.txt
astronomer-cosmos==1.0.4
apache-airflow-providers-postgres==5.6.0
python-dotenv
requests
pendulum
